Copied everything dto-related from [Grafana 6.2.5 documentation](https://grafana.com/docs/grafana/v6.2/reference/dashboard/) as precise as I could. Will need to figure out many details, though.
As for the architecture, I plan for dto to be immutable. At least I can copy with Kotlin's built-in `copyOf` and supply changed fields there. The plan is to have some kind of more high-level builders that are aware of other widgets/panels, queries, variables and such, and to have them space out the panels on the layout and to insert the queries as strings. Possibly with checking if the dashboard has the variables needed by the query. Another thing is to try to keep number of abstraction layers as thin as possible.